requirements.txt
mysql-connector-python
SQLAlchemy
python-dotenv
requests
bcrypt
PyJWT